some build fixes (barely)
